pyproject.toml
[tool.poetry]
name = "friend-server"
version = "0.1.0"
description = ""
authors = ["moreal <dev.moreal@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "0.78.0"
uvicorn = "^0.18.2"
requests = "^2.28.1"
fastapi-sessions = "^0.3.2"
python-dotenv = "^0.20.0"
databases = "^0.6.0"
strawberry-graphql = {extras = ["fastapi"], version = "^0.117.0"}
cryptography = "^37.0.4"
PyMySQL = "^1.0.2"
mysqlclient = "^2.1.1"
aiomysql = "^0.1.1"
orm = {extras = ["mysql"], version = "^0.3.1"}
ormar = "^0.11.2"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
mypy = "^0.961"
types-requests = "^2.28.0"
types-SQLAlchemy = "^1.4.42"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
